rename applet component
