+++ src/Components/Applets/AppletCard.vue
